requirements.txt
Flask==3.1.0
tensorflow==2.17.1
numpy==1.24.3
Pillow==9.5.0
h5py==3.8.0
gunicorn==20.1.0  # If deploying on a platform like Heroku
matplotlib==3.7.1
scikit-learn==1.2.2
Werkzeug==2.3.7  # Flask dependency
